Implemented a lower level method with timers.
The old implementation was renamed to phase_checker_simple,
the new method uses Timer1 to generate a 50 Hz reference
signal and the Input Capture function of the timer.
The clock divider is set to 8 so the interval is 500 ns.
One mains cycle is therefore approximately 40 000 timer
counts. While plugged in the frequency is constantly synced
to the input. I still need to find a way to keep a precise
clock when unplugged. The drift is now slow enough to compare
two outlets that are close to eachother, but ideally I want
to have the ability to walk to another building and find the
matching phase there so the clock needs to keep approximate
sync for maybe 10 minutes.