Implemented period averaging.
The period of the last 256 cycles is now stored
and upon disconnecting the mains, it is averaged
to produce a better approximation of the original
frequency. This works quite well when tested with
another MCU as a frequency source, the drift is
very low even over 10 minutes. However it turns out
that the grid frequency is anything but stable.
Variations of 10 mHz are perfectly acceptable for
the grid but it means going completely out of phase
in about 50 seconds.